+++ gin/sentrygin_test.go
@@ -330,7 +330,7 @@ func TestIntegration(t *testing.T) {
 			sentry.Event{},
 			"Contexts", "EventID", "Extra", "Platform", "Modules",
 			"Release", "Sdk", "ServerName", "Tags", "Timestamp",
-			"sdkMetaData",
+			"sdkMetaData", "attachments",
 		),
 		cmpopts.IgnoreFields(
 			sentry.Request{},
@@ -354,7 +354,7 @@ func TestIntegration(t *testing.T) {
 			sentry.Event{},
 			"Contexts", "EventID", "Platform", "Modules",
 			"Release", "Sdk", "ServerName", "Timestamp",
-			"sdkMetaData", "StartTime", "Spans",
+			"sdkMetaData", "StartTime", "Spans", "attachments",
 		),
 		cmpopts.IgnoreFields(
 			sentry.Request{},